feat(pvm): complete the riscv tests (#51)
* feat(codec): numberic encoding without trailing zeros in bytes

* feat(codec): introduce variable length encoding

* feat(pvm): get length on prefix decoding

* feat(pvm): correct the implementation of deblob

* feat(pvm): support initial pc

* feat(pvm): introduce branch

* feat(pvm): apply branch fn to branch instructions

* feat(pvm): use i32 for offsets

* chore(pvm): try adding missing visitors

* chore(pvm): update pc on jumping

* feat(pvm): impl display for instructions

* feat(pvm): correct memory write and dynamic jump

* feat(pvm): correct offset handler

* feat(pvm): correct mul upper functions

* feat(pvm): complete rotation operations

* feat(pvm): make clippy happy

+++ crates/pvm/parser/src/format/rio.rs
@@ -7,7 +7,7 @@ impl From<RIO> for Vec<u8> {
         let rii = RII {
             reg0: value.reg0,
             imm0: value.imm0,
-            imm1: value.off0,
+            imm1: value.off0 as u64,
         };
 
         rii.into()
@@ -24,7 +24,7 @@ impl From<&[u8]> for RIO {
         RIO {
             reg0: (bytes[0] % 16).min(12),
             imm0: u64::read_imm(&bytes[1..mid]),
-            off0: u64::read(&bytes[mid..]).0,
+            off0: u64::read(&bytes[mid..]).0 as i32,
         }
     }
 }
